+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added handling of incorrect layout paths via the **format** command.
 * Refactor **create-content-artifacts** command - Efficient artifacts creation and better logging.
 * Fixed an issue where image and description files were not handled correctly by **validate** and **update-release-notes** commands.
+* Fixed an issue where the **format** command didn't remove all extra fields in a file
 * Added a validation for tlp_color param in feeds in **validate** command.
 
 # 1.1.11